+++ sdk/resourcemanager/devopsinfrastructure/armdevopsinfrastructure/subscriptionusages_client_example_test.go
@@ -17,8 +17,8 @@ import (
 	"github.com/Azure/azure-sdk-for-go/sdk/resourcemanager/devopsinfrastructure/armdevopsinfrastructure"
 )
 
-// Generated from example definition: https://github.com/Azure/azure-rest-api-specs/blob/33c4457b1d13f83965f4fe3367dca4a6df898100/specification/devopsinfrastructure/resource-manager/Microsoft.DevOpsInfrastructure/preview/2024-04-04-preview/examples/SubscriptionUsages_ListByLocation.json
-func ExampleSubscriptionUsagesClient_NewListByLocationPager() {
+// Generated from example definition: https://github.com/Azure/azure-rest-api-specs/blob/ff7b8e12e78b352561e2e470dd045be310a313fa/specification/devopsinfrastructure/resource-manager/Microsoft.DevOpsInfrastructure/preview/2024-04-04-preview/examples/SubscriptionUsages_Usages.json
+func ExampleSubscriptionUsagesClient_NewUsagesPager() {
 	cred, err := azidentity.NewDefaultAzureCredential(nil)
 	if err != nil {
 		log.Fatalf("failed to obtain a credential: %v", err)
@@ -28,7 +28,7 @@ func ExampleSubscriptionUsagesClient_NewListByLocationPager() {
 	if err != nil {
 		log.Fatalf("failed to create client: %v", err)
 	}
-	pager := clientFactory.NewSubscriptionUsagesClient().NewListByLocationPager("eastus", nil)
+	pager := clientFactory.NewSubscriptionUsagesClient().NewUsagesPager("eastus", nil)
 	for pager.More() {
 		page, err := pager.NextPage(ctx)
 		if err != nil {
@@ -39,33 +39,27 @@ func ExampleSubscriptionUsagesClient_NewListByLocationPager() {
 			_ = v
 		}
 		// If the HTTP response code is 200 as defined in example definition, your page structure would look as follows. Please pay attention that all the values in the output are fake values for just demo purposes.
-		// page.QuotaListResult = armdevopsinfrastructure.QuotaListResult{
+		// page.PagedQuota = armdevopsinfrastructure.PagedQuota{
 		// 	Value: []*armdevopsinfrastructure.Quota{
 		// 		{
-		// 			Name: to.Ptr("standardDADSv5Family"),
-		// 			ID: to.Ptr("/subscriptions/00000000-0000-0000-0000-000000000000/providers/Microsoft.Compute/locations/eastus/providers/Microsoft.DevOpsInfrastructure/Usages/standardDADSv5Family"),
-		// 			Properties: &armdevopsinfrastructure.QuotaProperties{
-		// 				Name: &armdevopsinfrastructure.QuotaName{
-		// 					LocalizedValue: to.Ptr("Standard DADSv5 Family vCPUs (PME VMSS)"),
-		// 					Value: to.Ptr("standardDADSv5Family"),
-		// 				},
-		// 				CurrentValue: to.Ptr[int64](0),
-		// 				Limit: to.Ptr[int64](212),
-		// 				Unit: to.Ptr("Count"),
+		// 			Name: &armdevopsinfrastructure.QuotaName{
+		// 				LocalizedValue: to.Ptr("Standard DADSv5 Family vCPUs (PME VMSS)"),
+		// 				Value: to.Ptr("standardDADSv5Family"),
 		// 			},
+		// 			CurrentValue: to.Ptr[int64](0),
+		// 			ID: to.Ptr("/subscriptions/00000000-0000-0000-0000-000000000000/providers/Microsoft.Compute/locations/eastus/providers/Microsoft.DevOpsInfrastructure/Usages/standardDADSv5Family"),
+		// 			Limit: to.Ptr[int64](212),
+		// 			Unit: to.Ptr("Count"),
 		// 		},
 		// 		{
-		// 			Name: to.Ptr("standardDPLDSv5Family"),
-		// 			ID: to.Ptr("/subscriptions/00000000-0000-0000-0000-000000000000/providers/Microsoft.Compute/locations/eastus/providers/Microsoft.DevOpsInfrastructure/Usages/standardDPLDSv5Family"),
-		// 			Properties: &armdevopsinfrastructure.QuotaProperties{
-		// 				Name: &armdevopsinfrastructure.QuotaName{
-		// 					LocalizedValue: to.Ptr("Standard DPLDSv5 Family vCPUs (PME VMSS)"),
-		// 					Value: to.Ptr("standardDPLDSv5Family"),
-		// 				},
-		// 				CurrentValue: to.Ptr[int64](0),
-		// 				Limit: to.Ptr[int64](100),
-		// 				Unit: to.Ptr("Count"),
+		// 			Name: &armdevopsinfrastructure.QuotaName{
+		// 				LocalizedValue: to.Ptr("Standard DPLDSv5 Family vCPUs (PME VMSS)"),
+		// 				Value: to.Ptr("standardDPLDSv5Family"),
 		// 			},
+		// 			CurrentValue: to.Ptr[int64](0),
+		// 			ID: to.Ptr("/subscriptions/00000000-0000-0000-0000-000000000000/providers/Microsoft.Compute/locations/eastus/providers/Microsoft.DevOpsInfrastructure/Usages/standardDPLDSv5Family"),
+		// 			Limit: to.Ptr[int64](100),
+		// 			Unit: to.Ptr("Count"),
 		// 	}},
 		// }
 	}